refactor(scheduler): isolate schema migrations
