+++ infrastructure/scheduling/scheduler/claim_store.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import json
 import logging
 import sqlite3
-import time
 from collections.abc import Sequence
 from dataclasses import dataclass
 from datetime import UTC, datetime, timedelta
@@ -14,18 +13,12 @@ from typing import Any
 from uuid import uuid4
 
 from config.constants import OPENSRE_HOME_DIR
+from infrastructure.scheduling.scheduler.migrations import apply_migrations
 from infrastructure.scheduling.scheduler.types import DeliveryOutcome, TaskRun, TaskStatus
 
 logger = logging.getLogger(__name__)
 
 _DB_FILENAME = "scheduler.db"
-
-#: How long to keep retrying a column add while a competing process holds the
-#: write lock. Each attempt re-reads the schema first, so a winner's commit
-#: ends the loop immediately; this only bounds how long a *stuck* writer is
-#: tolerated before the real error surfaces.
-_MIGRATION_TIMEOUT_SECONDS = 30.0
-_MIGRATION_RETRY_DELAY_SECONDS = 0.1
 
 #: How far back to look for a run with readable per-target history before
 #: reporting none. Only a bound on work, not on correctness: exhausting it
@@ -33,24 +26,6 @@ _MIGRATION_RETRY_DELAY_SECONDS = 0.1
 #: than widening to every destination.
 _TARGETED_RUN_SCAN_LIMIT = 50
 _CLAIM_LEASE_SECONDS = 30 * 60
-_TASK_RUNS_SCHEMA = """
-    CREATE TABLE task_runs (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        task_id TEXT NOT NULL,
-        fire_time TEXT NOT NULL,
-        attempt INTEGER NOT NULL DEFAULT 1,
-        started_at TEXT NOT NULL,
-        finished_at TEXT,
-        status TEXT NOT NULL DEFAULT 'pending',
-        posted_message_id TEXT DEFAULT '',
-        error TEXT DEFAULT '',
-        provider TEXT DEFAULT '',
-        targets TEXT DEFAULT '',
-        owner_token TEXT NOT NULL DEFAULT '',
-        lease_expires_at TEXT NOT NULL DEFAULT '',
-        UNIQUE(task_id, fire_time, attempt)
-    )
-"""
 _RUN_COLUMNS = (
     "task_id, fire_time, started_at, finished_at, status, posted_message_id, "
     "error, provider, targets, attempt"
@@ -80,90 +55,12 @@ def _connect(db_path: Path) -> sqlite3.Connection:
     return conn
 
 
-def _ensure_schema(conn: sqlite3.Connection) -> None:
-    """Create or migrate task runs, rechecking under a write lock before changes."""
-    columns = _table_columns(conn)
-    if {"attempt", "targets"} <= columns:
-        return
-
-    conn.execute("BEGIN IMMEDIATE")
-    try:
-        columns = _table_columns(conn)
-        if not columns:
-            conn.execute(_TASK_RUNS_SCHEMA)
-        elif "attempt" not in columns:
-            _migrate_legacy_claim_table(conn, columns)
-        _add_missing_columns(conn)
-        conn.commit()
-    except Exception:
-        conn.rollback()
-        raise
-
-
-def _table_columns(conn: sqlite3.Connection, table: str = "task_runs") -> set[str]:
-    return {str(row[1]) for row in conn.execute(f"PRAGMA table_info({table})")}
-
-
-def _has_targets_column(conn: sqlite3.Connection) -> bool:
-    return "targets" in _table_columns(conn)
-
-
-def _migrate_legacy_claim_table(conn: sqlite3.Connection, legacy_columns: set[str]) -> None:
-    """Rebuild the pre-lease table so reclaimed attempts retain their history."""
-    conn.execute("ALTER TABLE task_runs RENAME TO task_runs_legacy")
-    conn.execute(_TASK_RUNS_SCHEMA)
-    targets = "targets" if "targets" in legacy_columns else "''"
-    conn.execute(
-        "INSERT INTO task_runs "
-        "(id, task_id, fire_time, attempt, started_at, finished_at, status, "
-        "posted_message_id, error, provider, targets, owner_token, lease_expires_at) "
-        f"SELECT id, task_id, fire_time, 1, started_at, finished_at, status, "
-        f"posted_message_id, error, provider, {targets}, '', started_at "
-        "FROM task_runs_legacy"
-    )
-    conn.execute("DROP TABLE task_runs_legacy")
-
-
-def _add_missing_columns(conn: sqlite3.Connection) -> None:
-    """Add columns introduced after a database was first created.
-
-    Two processes can both see the column missing before either commits its
-    ``ALTER TABLE``, so the loser's own write fails — with "duplicate column"
-    if the winner already committed, or with a lock-timeout if the winner is
-    still in flight and outlasts ``busy_timeout``. Rechecking the schema
-    (rather than matching the error text) covers the first case, and retrying
-    covers the second: at timeout the winner's column is not visible yet, so
-    a single recheck would wrongly conclude the migration failed. Each retry
-    re-reads the schema first, so the loser returns the moment the winner's
-    commit lands.
-
-    Adding a column to this table is sub-millisecond work, so a writer still
-    holding the lock after ``_MIGRATION_TIMEOUT_SECONDS`` is stuck rather than
-    slow. Raising then is deliberate: retrying forever would hide a wedged
-    database behind a scheduler that never fires.
-    """
-    deadline = time.monotonic() + _MIGRATION_TIMEOUT_SECONDS
-    while True:
-        if _has_targets_column(conn):
-            return
-        try:
-            conn.execute("ALTER TABLE task_runs ADD COLUMN targets TEXT DEFAULT ''")
-        except sqlite3.OperationalError:
-            if _has_targets_column(conn):
-                return
-            if time.monotonic() >= deadline:
-                raise
-            time.sleep(_MIGRATION_RETRY_DELAY_SECONDS)
-        else:
-            return
-
-
 def try_claim(task_id: str, fire_time: str, db_path: Path | None = None) -> ExecutionClaim | None:
     """Claim a task tick or reclaim it when the current lease has expired."""
     path = db_path or _default_db_path()
     conn = _connect(path)
     try:
-        _ensure_schema(conn)
+        apply_migrations(conn)
         conn.execute("BEGIN IMMEDIATE")
         now = datetime.now(UTC)
         now_text = now.isoformat()
@@ -240,7 +137,7 @@ def complete_run(
     path = db_path or _default_db_path()
     conn = _connect(path)
     try:
-        _ensure_schema(conn)
+        apply_migrations(conn)
         now = datetime.now(UTC).isoformat()
         cursor = conn.execute(
             "UPDATE task_runs SET finished_at = ?, status = ?, "
@@ -325,7 +222,7 @@ def get_runs(task_id: str, limit: int = 20, db_path: Path | None = None) -> list
     path = db_path or _default_db_path()
     conn = _connect(path)
     try:
-        _ensure_schema(conn)
+        apply_migrations(conn)
         cursor = conn.execute(
             f"SELECT {_RUN_COLUMNS} "
             "FROM task_runs WHERE task_id = ? ORDER BY started_at DESC LIMIT ?",
@@ -345,7 +242,7 @@ def get_latest_finished_run(task_id: str, db_path: Path | None = None) -> TaskRu
     path = db_path or _default_db_path()
     conn = _connect(path)
     try:
-        _ensure_schema(conn)
+        apply_migrations(conn)
         cursor = conn.execute(
             f"SELECT {_RUN_COLUMNS} "
             "FROM task_runs WHERE task_id = ? AND status IN (?, ?) "
@@ -375,7 +272,7 @@ def get_latest_targeted_run(task_id: str, db_path: Path | None = None) -> TaskRu
     path = db_path or _default_db_path()
     conn = _connect(path)
     try:
-        _ensure_schema(conn)
+        apply_migrations(conn)
         cursor = conn.execute(
             f"SELECT {_RUN_COLUMNS} "
             "FROM task_runs WHERE task_id = ? AND status IN (?, ?) "
@@ -408,7 +305,7 @@ def delete_runs(task_id: str, db_path: Path | None = None) -> int:
         return 0
     conn = _connect(path)
     try:
-        _ensure_schema(conn)
+        apply_migrations(conn)
         cursor = conn.execute(
             "DELETE FROM task_runs WHERE task_id = ?",
             (task_id,),
