+++ tests/scheduler/test_claim_store.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import inspect
 import sqlite3
 import threading
 from concurrent.futures import ThreadPoolExecutor
@@ -9,7 +10,7 @@ from pathlib import Path
 
 import pytest
 
-from infrastructure.scheduling.scheduler import claim_store
+from infrastructure.scheduling.scheduler import claim_store, migrations
 from infrastructure.scheduling.scheduler.claim_store import (
     ExecutionClaim,
     complete_run,
@@ -59,6 +60,15 @@ class _AlterFailsConnection:
         if sql.strip().startswith("ALTER TABLE"):
             raise sqlite3.OperationalError("database is locked")
         return self._conn.execute(sql, *args)
+
+
+def test_schema_ddl_is_owned_by_the_migration_module() -> None:
+    """Keep runtime claim queries separate from schema evolution."""
+    claim_store_source = inspect.getsource(claim_store).upper()
+
+    assert "CREATE TABLE" not in claim_store_source
+    assert "ALTER TABLE" not in claim_store_source
+    assert "DROP TABLE" not in claim_store_source
 
 
 class TestClaimStore:
@@ -182,7 +192,7 @@ class TestClaimStore:
         # start-ordered lookback of five would drop the finished row.
         conn = claim_store._connect(db_path)
         try:
-            claim_store._ensure_schema(conn)
+            migrations.apply_migrations(conn)
             conn.execute(
                 "INSERT INTO task_runs "
                 "(task_id, fire_time, started_at, finished_at, status) "
@@ -281,7 +291,7 @@ class TestConcurrency:
         conn.commit()
         conn.close()
 
-        real_table_columns = claim_store._table_columns
+        real_table_columns = migrations._table_columns
         unlocked_reads = threading.Barrier(2)
 
         def _synchronize_unlocked_reads(
@@ -293,7 +303,7 @@ class TestConcurrency:
                 unlocked_reads.wait(timeout=5)
             return columns
 
-        real_migrate = claim_store._migrate_legacy_claim_table
+        real_migrate = migrations._migrate_legacy_claim_table
         migration_count = 0
         count_lock = threading.Lock()
 
@@ -306,8 +316,8 @@ class TestConcurrency:
                 migration_count += 1
             real_migrate(migration_conn, columns)
 
-        monkeypatch.setattr(claim_store, "_table_columns", _synchronize_unlocked_reads)
-        monkeypatch.setattr(claim_store, "_migrate_legacy_claim_table", _count_migration)
+        monkeypatch.setattr(migrations, "_table_columns", _synchronize_unlocked_reads)
+        monkeypatch.setattr(migrations, "_migrate_legacy_claim_table", _count_migration)
 
         with ThreadPoolExecutor(max_workers=2) as pool:
             claims = list(
@@ -450,7 +460,7 @@ class TestPerTargetOutcomes:
         conn.execute("ALTER TABLE task_runs ADD COLUMN targets TEXT DEFAULT ''")
         conn.commit()
 
-        real_has_column = claim_store._has_targets_column
+        real_has_column = migrations._has_targets_column
         calls = {"count": 0}
 
         def _lie_on_first_call(check_conn: sqlite3.Connection) -> bool:
@@ -460,9 +470,9 @@ class TestPerTargetOutcomes:
             # nothing about the recovery path.
             return False if calls["count"] == 1 else real_has_column(check_conn)
 
-        monkeypatch.setattr(claim_store, "_has_targets_column", _lie_on_first_call)
+        monkeypatch.setattr(migrations, "_has_targets_column", _lie_on_first_call)
 
-        claim_store._add_missing_columns(conn)  # must not raise
+        migrations._add_missing_columns(conn)  # must not raise
         assert calls["count"] == 2
 
     def test_a_winner_committing_after_our_timeout_is_picked_up_on_retry(
@@ -501,9 +511,9 @@ class TestPerTargetOutcomes:
             checks["count"] += 1
             return checks["count"] > 2
 
-        monkeypatch.setattr(claim_store, "_has_targets_column", _column_appears_on_the_retry)
+        monkeypatch.setattr(migrations, "_has_targets_column", _column_appears_on_the_retry)
 
-        claim_store._add_missing_columns(_AlterFailsConnection(conn))  # must not raise
+        migrations._add_missing_columns(_AlterFailsConnection(conn))  # must not raise
 
         # Pre-check, post-timeout recheck, then the retry's own pre-check.
         assert checks["count"] == 3
@@ -513,7 +523,7 @@ class TestPerTargetOutcomes:
     ) -> None:
         """A writer stuck past the budget surfaces the real error rather than
         retrying forever."""
-        monkeypatch.setattr(claim_store, "_MIGRATION_TIMEOUT_SECONDS", 0.2)
+        monkeypatch.setattr(migrations, "_MIGRATION_TIMEOUT_SECONDS", 0.2)
         conn = sqlite3.connect(str(db_path))
         conn.execute("""
             CREATE TABLE task_runs (
@@ -532,7 +542,7 @@ class TestPerTargetOutcomes:
         conn.commit()
 
         with pytest.raises(sqlite3.OperationalError, match="database is locked"):
-            claim_store._add_missing_columns(_AlterFailsConnection(conn))
+            migrations._add_missing_columns(_AlterFailsConnection(conn))
 
 
 class TestLatestTargetedRun:
